server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pyautogui>=0.9.54
mss>=9.0.1
pyperclip>=1.8.2
Pillow>=10.0.0
